feat(IPX): add atomic-ipx-recs-list (#3966)
For IPX, we need to add a pageView event in the actions history when we
click on the grid containing an result list item.

To do that, I copied the atomic-recs-list component and made a
atomic-ipx-recs-list component. Its exactly the same except that it
modifies the onSelect method of the built interactiveResult.

Another way to go would be to modify the origin atomic-recs-list to
accept custom buildRecsInteractiveResult methods but its not the way we
usually do for IPX components (as usually we dont modify origin atomic
components to fit the needs of the experimental IPX components)

Tested locally ✅

Note: We already have a mechanism to add `pageView` events to
`actionsHistory` for search and result items. We were just missing the
same mechanism when the click is on the container itself
(`result-list-grid-clickable-container`) rather than the item link