feat(Commerce Headless): change the way selectors are exposed (#4016)
The initial idea was to refactor all selectors in a way that the
arguments of the selector function would be the commerce engine rather
than the state.

However, this would require a significant amount of effort for the
current need. Since we only need to use a couple of selectors from all
the existing ones (and it is unlikely that we will need to use the
majority of them outside of headless), it is not wise IMO to spend all
the time changing all the selectors and their unit tests.

Instead, I have simply wrapped all the selectors that we need outside of
headless in a function that accepts an engine.

WDYT?

https://coveord.atlassian.net/browse/KIT-3199

---------

Co-authored-by: Olivier Lamothe <olamothe@coveo.com>
Co-authored-by: GitHub Actions Bot <>
Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>
Co-authored-by: Louis Bompart <lbompart@coveo.com>
Co-authored-by: Felix Perron-Brault <fpbrault@coveo.com>