chore: add deprecation note for DNT (#4020)
The PR adds a deprecation note to remove DNT as part of the next major
release.
The decision is captured
[here](https://coveord.atlassian.net/browse/LENS-1502?focusedCommentId=764668).
The main drivers were:
- DNT is an outdated standard. Newer standards like GPC have been
introduced. We don't believe it's Coveo's role to update libraries to
adhere to the latest standards?
- Privacy standards differ around the world. Baking a single privacy
stance into Coveo's libraries and apis will be overly restrictive in
certain geographies.

**Aside**
I tried to adjust the logic in headless to only apply DNT if the
analyticsMode is `legacy`. However, this required changing an action to
an thunk. The signature change created multiple typescript errors, and
so seems like it's breaking. I changed course to add a deprecation note
instead. With the v3 release around the corner, it seems less risky to
simply remove the `doNotTrack` function rather than making a signature
change.

---------

Co-authored-by: Louis Bompart <lbompart@coveo.com>