chore: fix several release issues (#4059)
A good bundle of 3 issues.

## The NXpocalypse

Originally, my removal of the plugin caused that, but I still think the
plugin was a bit of a code-smell hiding some other issue, such as the
following one.
For some obscure reason, I did not succeed in deciphering
`release:phase2`(before my changes), which systematically refuses to run
its dependents, even when they are defined.
So, I scrapped it and replaced it with a single task, `release:phase1`
(after my changes). It does the same thing, but it's a bit simpler.

## Whoopsie daisy in
https://github.com/coveo/semantic-monorepo-tools/commit/916471193fafd59c82472af562a91dd1042b35d8

So, fun fact: The command works as intended on Windows and returns the
correct stuff.
On Linux/macOS, the double quotes are not interpreted nicely or the same
way, and every commit ends with `format:`.
That ain't a good semantic commit message, so since we updated
https://github.com/coveo/ui-kit/commit/eb3542e592888dd14196bbcbec230a58cb6e7d86,
every release has been a patch...

Will be fixed in coveo/semantic-monorepo-tools#240, but we can remove
the patch when we update, it'll be about the same.

## Reify choke on packages installing themselves as a dependency.

It's probably a way to make it work, but I decided to replace the
version by ' file:.`. It's a dev dep, so it has no impact on the
consumers and saves me some headaches.

## Was it tested?

Yes: I used some admin magic to be able to run the Prerelease enviro on
this branch until it worked. This is now revoked/back to normal.