feat(headless,atomic): fix issues for commerce did you mean + handle query trigger for commerce (#4053)
Add a new `search-actions-thunk-processor` file for commerce search
actions (which mimics what we have for search use case)

It is in charge of processing the search response specifically (not
product listing) since there can be a mismatch between what Headless
think was the query executed (what the user submitted in the search box)
and what was actually executed in the backend (query correction/query
trigger)

When there is a mismatch, we need to reconcile the search boxes (query
set slice) and the query itself (query slice) by dispatching actions.
Otherwise it causes bug where the end user will see that did you mean
component says "query `foo` was executed" but the search box will
display `bar`.

Since it can be a complicated piece of code to push into the thunk
function itself, the `processor` file is there to isolate and make it
easier to test the behaviour.


https://coveord.atlassian.net/browse/KIT-3224

---------

Co-authored-by: GitHub Actions Bot <>