feat(commerce): mark v2 listings as non-v2 and non-v2 listings as old (#4087)
This is somewhat of an overdue rename. Generally, I've renamed
"ProductListingV2" to "ProductListing", and the v1 listings from
"ProductListing" to "OldProductListing".

The renamed classes or file names do not affect the API surface of
neither the product-listing sub-package, or the commerce sub-package.
I've however deleted the v2 facet analytics actions, and removed them
from the `product-listing-actions-loader.ts` action loader, which is a
breaking change. I don't expect anyone to have been using these though.

[CAPI-763]

[CAPI-763]:
https://coveord.atlassian.net/browse/CAPI-763?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ