fix(headless): update product interface to reflect how the commerce api behaves (#4104)
This is less changes that what I had originally thought.

The `BaseProduct` interface should now properly reflect what's returned
by the commerce API, following the source of truth (the code of the
service itself).

* Most of the change are about changing from `foo | undefined` to `foo |
null`.
* Array values are never null, but might be an empty array (due to how
the service is coded)
* ec_category is always an array, not a string
* Added couple missing properties (non breaking)

There is only "one lie" in this change: `excerpts, nameHighlights,
excerptsHighlights` are "always defined when in a search use case", and
"always undefined when in a listing or recommendation use case".

I thought about splitting and creating multiple interface
`SearchProduct, ListingProduct, RecommendationProduct`, and remove
`Product` as it exists right now.

However this would be a very noisy change, for existing deployments.
`Product` is an *important* interface, which is nearly always referenced
by any implementer using TypeScript.

So I made the judgment call to "lie a little bit, for the greater good"
😂 They are marked as possibly undefined for all use cases.

https://coveord.atlassian.net/browse/KIT-3164