fix(atomic): fix atomic-commerce-search-box needlessly requesting suggestions when input is disabled (#4096)
Calling `searchBox.updateText()` triggers a query suggestion request in
Headless: This is fine, and generally what implementers requires the
search box to do.

However, in atomic, we've "split" the suggestion outside the searchbox
component itself into a `SuggestionManager` utility class, which should
be the one in charge of requesting suggestions when necessary.

To fix, exported `loadQuerySetActions` from Headless commerce, and use
this action directly to update the state of the search box, without
requesting `/querySuggest`.

https://coveord.atlassian.net/browse/KIT-3320

---------

Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>