fix(commerce): populate `context.user.userAgent` with `navigatorContext` (#4113)
headless was populating the `context.user.userAgent` property with what
was set through the `setUser`. With this PR, we extract the `userAgent`
and the `referrer` from the relay instance and use these values instead.
This makes using the headless commerce controllers less error-prone.

⚠️ This means we can remove the `setUser` action (and its sibling method
on the context controller), and `setView`'s `referrer` parameter, which
I do in this PR. Do you think this makes sense? Is there a use case for
allowing manual control over these parameters? @samisayegh
@louis-bompart @fbeaudoincoveo ?

[CAPI-1091]

[CAPI-1091]:
https://coveord.atlassian.net/browse/CAPI-1091?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>
Co-authored-by: Louis Bompart <lbompart@coveo.com>