fix(commerce): expose child product directly on `promoteChildToParent` (#4128)
Instead of mutating the received products from the API, I propose we
dynamically find the corresponding parent and children in the reducer.
This _seems_ less error-prone to me, because we don't need to mutate
products in every place we fulfill a request containing `products`, at
the cost of being less performant, as we iterate on every product every
time a children is promoted. Thoughts?

An other option to consider would be to mutate the products directly in
the commerce api client.

https://coveord.atlassian.net/browse/KIT-3221