feat(knowledgeAPI): api slice (#4014)
## Overview
This update introduces a client to manage the new AnswerApi stream,
serving RGA model answers. The client leverages [Redux Toolkit
Query](https://redux-toolkit.js.org/rtk-query/overview).

## Tech
`baseKnowledgeAPISlice.ts`
This file contains the baseQuery, which can be used for simple API
implementations, and the main slice definition, which can be used to
create the Redux entry in the state.

`answerApiSlice.ts`
This file defines the endpoint that will manage the streaming,
completing the Redux API slice.

While the `dynamicBaseQuery` pattern sets up a basic structure for
adding new knowledge endpoints in the future, the code-splitting pattern
is useful for maintaining a separation of concerns between generic code
and endpoint-specific code.

## The Feature
This feature creates an entry in the state, forming the foundation of
the new answer controller for the answer generation stream. It manages
everything that is currently manually dispatched regarding API-level
activity, such as `isStreaming`, `isLoading`, `answer`, and `citations`.

The RTK/Query library already handles the multi-query pattern,
eliminating the need for complex ID management through a custom random
ID implementation.


![image](https://github.com/coveo/ui-kit/assets/45688129/fcb4d6ce-f810-4937-bd5f-befd92d03d22)

**Edit** Now works with the new AnswerApi endpoint

![image](https://github.com/coveo/ui-kit/assets/45688129/bdbf1be2-a076-4dba-875f-d15fc4f1c5f2)

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>