test(atomic): add atomic-commerce-facet tests (#4123)
Add tests for `atomic-commerce-facet`.

You will see some shenanigan in story setup to "keep only a single
facet".

As you know, atomic-commerce-facet are generated/internal: So we need to
trick storybook a bit to output `atomic-commerce-facets` (with an `s`),
and then "snipe" components that are not of the correct type (ie:
everything that is not an `atomic-commerce-facet`).

This works pretty well for the test themselves. We get a page with only
a specific type of facet, and we can run test as normal.

There will be work needed still to make storybook work correctly with
the Shadow parts.

The shadow parts are properly documented; It's just that when they get
customized in Storybook UI, the styling output is not reflected
correctly in the story.

However, I prefer to continue with testing and revisit that at a later
time.

https://coveord.atlassian.net/browse/KIT-3257

---------

Co-authored-by: GitHub Actions Bot <>