feat(commerce): distinguish facet search solution type (#4156)
To allow the facet search endpoint to resolve either the listing or
search configuration on a facet search, we pass in a `type` query param
that's either `SEARCH` or `LISTING`.

Since the search configuration also powers field suggestions, `SEARCH`
is also used. In the future, if we change field suggestions to be
powered by their own configuration, we'll have to change this parameter.

search call:

![image](https://github.com/coveo/ui-kit/assets/8978908/45380e37-a9c9-41e8-9a9e-f99607646825)

listing call:

![image](https://github.com/coveo/ui-kit/assets/8978908/2fbb99f4-4b90-4628-b2cc-48b92fd44788)


[CAPI-911]

[CAPI-911]:
https://coveord.atlassian.net/browse/CAPI-911?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ