feat(answer): knowledge controller preparation (#4129)
## Summary

This change is just moving logic around; this increment is not modifying
any functionalities.

The change intents to prepare the addition of a new answer controller
that will leverage the new answerApi client. The end goal is to separate
the legacy logic of using the streamId from the searchApi, from the new
logic of the answerApi stream.

## Why is this needed

Since both logics (search-api VS answer-api) share a strong common root,
the current change will permit us to reuse the current
headless-core-generated-answer controller in both cases.

## Testing

Those change must be tested to ensure that the current functionalities
stay unchanged

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>