feat(answer): error handling (#4146)
## Summary

Updates to the Answer Api client to make sure the error handling is
proper. Some little extras needed for it to work proper that will make
the subsequent review a bit easier.

## Why

The Answer Api is almost just a proxy for the stream coming from reveal.
And for some technical reasons, the call to the answer api will return
an error message but will not return downright error status codes like
400s when the stream fails.
That said, we can rely on the stream end message in case the stream is
not working properly, and since its better to give the user or the devs
all the information we can, this increment makes sure that the error
will appear in the logs and in the state in case we want to leverage the
error message in the front end component.

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>