feat(answer): knowledge answer controller (#4135)
## Summary

This increment adds the Controller that will map the state of the
answerApi Slice to the Generated Answer State used by the front end
components.

## Why this is needed

Since we must for a while support the current working flow with the API
stream, while adding the new flow from the AnswerAPI, it was way simpler
to make sure both implementation rely on the same state. Cuz having to
rework all the components to align them with the AnswerApi slice
directly would be breaking and would not cover all the actual
functionalities, like rephrasing.

## How does it work

The front end components will use the same controller in both the flows,
the only thing changing, at a DX level, is to pass the configuration Id
in the props of the controller.
By doing so, the first level controller will then build and extend the
right sub controller to handle the proper GenAI flow.
Both the legacy controller working with the searchUi and the new
controller working with the Answer API share the same state.
The configuration Id will be passed to the client seamlessly. 
In most cases, the answer generation will be identical in both the
flows. The main difference is that the flow using the
answerConfigurationId unlocks the answer management capabilities.

---------

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>