feat(headless): trigger request answer from search request (#4148)
## Summary

Addition of a listener that will call the answer API when a new search
request ID is detected.

## Why is it needed

The AnswerApi stream will not be automatically handled by the search
like the current flow is. The api must be queried explicitly.

## How it works

When a request ID is updated in the state by the regular search pattern,
the knowledge answer api will trigger a call for a stream with the
query.

[simplescreenrecorder-2024-07-15_16.38.46.webm](https://github.com/user-attachments/assets/f7316535-9307-4943-bcb1-156f33adefbf)

---------

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>