fix(headless): separate manual numeric facet in a separate slice for commerce (#4183)
This PR should fix various issues that were reported by QA surrounding
"manual numeric facets"; ie: the input in Atomic that allows end user to
select an arbitrary range. The issues were linked to the fact that we
tied together "the ranges returned by the API/computed by the index"
with those entered "manually" by the end user.

This PR creates a new slice ( `manual-numeric-facet-slice` ) and record
those entry in the URL as `mnf` (for manual numeric facet).

It also removes the `range` stateful property in Atomic, to instead rely
on a stateful value provided by the Headless controller itself.



KIT-3240

---------

Co-authored-by: GitHub Actions Bot <>