feat(quantic): new feedback modal for RGA (#4162)
Mockups:
https://www.figma.com/design/l63i3ZXWrVxldOR4WUBKS1/CRGA-Component?node-id=5514-14647&t=pOZ5lDKywVYJCqgI-0

SFINT-5555:

New Quantic RGA Feedback modal


![image](https://github.com/user-attachments/assets/4ffd312c-e0ce-475d-87f8-17cbdb288fc7)

The top 4 questions are required, and you have an example of how we tell
the user when they are required and missing.

The bottom two text fields are optional.

When you press send feedback there will be the same screen as before
thanking you for your feedback.

If you press skip it closes the modal. And nothing you've selected is
"remembered" if you re-open it.

The modal now opens on both like and dislike (thumbs up and thumbs
down), and as part of the event we send, we specify if the feedback was
positive or negative. We still send a separate event already when you
press thumbs up or down.

You can "change your mind", even if you send feedback, you can still
change your thumbs up to a thumbs down. but it will NOT re-open the
modal if you've already submitted feedback once.
Also if you click a 2nd time on thumbs down for example we won't re-open
the modal.

Everything is tested through unit tests since this is only the "content"
of the modal, and not the feedback process itself that we're testing.