fix(commerce): properly memoize suggestions controllers (#4184)
Field suggestions controllers weren't being updated because we were
memoizing their generation only on the list of facets, which did not
change when field suggestions values changed. Now, we include the facet
set (which includes facet values), to the function on which to memoize.
This ensures the field suggestions controllers are update with the right
values.

[CAPI-1102]

[CAPI-1102]:
https://coveord.atlassian.net/browse/CAPI-1102?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ