feat(answerApi): evaluations (#4239)
# Answer Api


[simplescreenrecorder-2024-07-30_16.55.09.webm](https://github.com/user-attachments/assets/79153f5a-19e6-42ec-8a3b-1c3921783e81)

## Summary

This PR make it so we can send CRGA evaluations from the feedback modal
to the AnswerApi

- The endpoint that we can POST evaluations to has been added
- The Steam answer client will get the header of the stream response and
set the Answer ID in its state
- The Controller now overrides the `SendFeedback` method in order to
POST the feedback to the answer api

### Post Evaluation Endpoint

A RTK query Endpoint have been added to the Answer slice
A little modification to the dynamicBaseQuery was beneficial to make the
endpoint definition simpler.

### Answer API Stream Headers recuperation

We use the `x-answer-id` from the stream call to identify the feedback
when we send it to the evaluation endpoint.

### sendFeedback method override

The native `sendFeedback` method from the base controller has been
modified to use the evaluation endpoint of the answerApi

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>