chore(headless SSR): add commerce ssr samples (#4228)
This PR introduces the headless-ssr-commerce sample project, utilizing
the @coveo/headless/ssr-commerce package. The project serves as a demo
application, currently featuring a product listing page. Although a
search feature is planned, it is not yet implemented.

Additionally, I’ve created and exposed two commerce controllers: summary
and product list. These controllers are included to provide content for
the listing page, but they are subject to change in the near future.

Here is what you will get at this stage after going to
http://localhost:3000/listing

![image](https://github.com/user-attachments/assets/73aebab6-3b38-412b-ae09-acdd49bc4853)

https://coveord.atlassian.net/browse/KIT-3441

---------

Co-authored-by: Alex Prudhomme <78121423+alexprudhomme@users.noreply.github.com>