fix(commerce): fix field suggestions state update (#4245)
Field suggestion values weren't updating when the query changed. To fix
this, we fetch the facet search state using a selector, just like we do
for facets (thanks @fbeaudoincoveo!).

I've also fixed a few other things:
- Load the proper reducers in field suggestions controllers
- Add the reducer to the commerce engine
- Type the reducer loaders
- Replace the `Subscribable` type to `Controller` on field suggestions
controllers
- Adjusted the field suggestions integration test to rely on controllers
instead of engine state updates. This was actually a clue that the
current behavior was broken, as it was not possible to await state
changes on the field suggestions controllers 😅

[CAPI-1201]

[CAPI-1201]:
https://coveord.atlassian.net/browse/CAPI-1201?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ