feat(headless SSR): support navigator context in both Engine and fetch/hydrate functions (#4231)
Introducing a method for configuring the navigator context provider.
Users are alerted with a warning if the navigator context is not
established on both the client and server sides.

I also updated the samples to demonstrate the navigator's application
within a Next.js framework, though the approach is not exclusive to this
framework.

There are two methods for setting the navigator context:
1. During engine configuration.
2. While initializing the app's static state and hydrating the app
state, if the first method is impractical.

The example in the SSR samples highlights a scenario where, due to the
framework's limitations, setting the navigator context directly within
the engine configuration is not feasible because headers are not yet
available. In such cases, the second method is recommended. However,
this approach necessitates invoking `setNavigatorContextProvider` on
both the server and client sides, increasing the potential for user
errors. To address this, a warning is issued if the setup is not
correctly implemented.