fix(commerce): namespace field suggestions to prevent clash with facet search (#4247)
Since facets in a global search configuration are used to power both
field suggestions and facets, it's possible that field suggestions for a
field be active at the same time that a facet search for the same field
is. This means that updating the text on a field suggestions controller
would cause a facet's facet search values to be updated as well. This is
because both the field suggestions and facet search were powered by the
same state.

To prevent this issue, I namespace the `facetId` used for field
suggestions **only** on the facet search reducers (the regular, and
category one).

I opted for this approach instead of reversing the dependency on the
facet search controller to read state from selectors since it would've
lead to much more changes in the controllers.

[CAPI-1201]

[CAPI-1201]:
https://coveord.atlassian.net/browse/CAPI-1201?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ