fix(answerStream): handle end of stream (#4270)
[SVCC-4057](https://coveord.atlassian.net/browse/SVCC-4057)

## Problem

The end of stream was no handled after a 'sorry' message.
The `answerGenerated` is false with the 'sorry' message, and this was
the condition to handle the end of stream.

The effect of this bug was that the cursor was not removed when the
answer was 'sorry'. Plus, the buttons to rephrase weren't showing.
Removing the possibility to rephrase.

## Solution

We now check if there is an answer that was outputed beside the official
`answerGenerated`.

## Side fix

We now send the `contentFormat` in the payload, like we do with the
SearchApi
 

[SVCC-4057]:
https://coveord.atlassian.net/browse/SVCC-4057?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>