fix(rga): state reset keep configid (#4286)
Fixing a regression when resetting the answer in the new answerAPI flow.

## Problem

When we need to reset a RGA answer
e.g. right after a new query has been submitted
The `answerConfigurationId` must be kept in the state whenever the
answer API flow is used.

## The Fix
We pass the actual state answerConfigurationID to the new state created
by the reset action.

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>