fix(answerApi): search context trigger new request (#4293)
## The Problem

With the actual search API flow. We make sure to retrigger a request for
a RGA answer when the search context changes.
e.g. Changing a facet.

We did not port this behavior when using the AnswerApi flow at first

## The solution

Since the AnswerApi generate answer call is basically proxying the
searchApi call. We can pass the params that should influence the search
context to the call, retriggering a request.

## Demo


[simplescreenrecorder-2024-08-19_08.47.49.webm](https://github.com/user-attachments/assets/52a45636-82fe-4763-9b7d-b8182d45b6a5)

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>