feat(headless): Add locale to insight search (#4302)
[SFINT-5668]

Essentially, you can now set the `locale` parameter in the
initialization options of the InsightEngine:

```
search: {
  locale: LOCALE,
}
```

It is under `search` because this is a concept that already existed on
the search engine that I am just bringing to the Insight Engine because
we need it.

It will be sent now as the `locale` parameter on the insight search
requests.


![image](https://github.com/user-attachments/assets/3dd809ee-36e0-4025-84d6-41f55668dbde)

I am re-using the existing Search Engine configuration action
`engine.dispatch(updateSearchConfiguration(search));` to update the
`engine.state.search.locale` value in the Insight Engine.


[SFINT-5668]:
https://coveord.atlassian.net/browse/SFINT-5668?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ