fix(atomic-react)!: remove analytics from atomic search interface (#4348)
In the atomic-react library, there are two methods to enable/disable
analytics:

```ts
buildSearchEngine({
  configuration: {
    ...getConfigurationForSample(sample),
    analytics: {enabled: false},
  },
});
```

Via AtomicSearchInterface:

```html
<AtomicSearchInterface analytics={false} />
```

### Issue
The second method, using AtomicSearchInterface, does not work as
expected. This is because the SearchInterfaceWrapper in atomic-react
reads the analytics configuration directly from the engine, ignoring the
analytics prop passed to AtomicSearchInterface.

### Solution
Remove analytics prop from the SearchInterfaceWrapper

https://coveord.atlassian.net/browse/KIT-3495