fix(quantic): Standalone searchbox dropdown overlaps with integrated searchbox in HSP (#4359)
[SFINT-5690](https://coveord.atlassian.net/browse/SFINT-5690)

### ISSUE:
- When you have a Standalone Searchbox and integrated searchbox at the
same time, the dropdown of the SSB overlaps with the integrated search
box. It was happening when the searchbox was expandable.

ie: 
<img width="1050" alt="image"
src="https://github.com/user-attachments/assets/cf7e8c60-7b2c-4a2c-92d2-0f0cbb1a15b0">


### THE CAUSE:

We have a `z-index-high` class on the wrapper of the
`quanticSearchboxInput` component, which used to set its z-index to 20 .

This was allowing the expandable searchbox to expand properly over the
page and still work with genQA tooltips (since those we set a z-index of
40).

I also noticed that a limitation was the class : `.cSearchPublisher` set
by Salesforce which sets a z-index of 10 to the salesforce wrapper of
the Hosted Standalone Searchbox. This we can't change unfortunately...
and since it was wrapping the SSB, it seems like it was causing the
issue since it was smaller than the 20 we tried to set. I know this
because if I change it in the dev console to 21 and above, it works
fine.


### SOLUTION:

I tried a few solutions such was wrapping the `standaloneSearchbox` in a
div with a higher z-index than 20, but it wasn't working because the
position of that element in the DOM was static and therefore would not
set a z-index properly.

Final solution: I reduced the `z-index-high` class z-index to 5 from 20.
This allows to work since its below the 10 of the `.cSearchPublisher`
and will still allow for the tooltips in RGA to fully be displayed as
their z-index is 40.

I know we agreed in another PR to z-index increments of 20, but this is
the best/most simple solution I have. We could also go by increments of
5?



### DEMO:


https://github.com/user-attachments/assets/ad433968-ceaf-4b68-8fc8-889941a267a2





[SFINT-5690]:
https://coveord.atlassian.net/browse/SFINT-5690?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ